requirements.txt
fastapi
uvicorn
pymupdf==1.25.3
python-docx==1.1.2
python-multipart
PyPDF2==3.0.1